Issue #5: Improve the Utility of the Trip Problem Report Admin Interface
- Add mouseover effect for the vehicle location records table, such that
the currently selected record will have its vehicle location highlighted
on the map.
- Add a row index column to the vehicle location records table, such
that it's easier to identify a specific record.
- For the info-window that's shown when you click on a vehicle
location record on the map, now show the record row index and a better
formatted timestamp, both of which should exactly match the values used
in the records table.