Add support for GTFS-realtime export to the onebusaway-api-webapp module.  This will expose the following API endpoints:
  /api/gtfs_realtime/alerts-for-agency/1.pb
  /api/gtfs_realtime/trip-updates-for-agency/1.pb
  /api/gtfs_realtime/vehicle-positions-for-agency/1.pb

which will return binary Protocol Buffer data in the GTFS-realtime schema.  The caller can optionally change the extension to .pbtext to see a textual debug representation of the message.

These methods are implemented with existing TransitDataService methods and can be thought of as a GTFS-realtime version of existing API methods like vehicles-for-agency.