Added headers to update urls
- Added a headers property to GtfsRealtimeSource class
- We can add header values through the data-source.xml file in
onebusaway-transit-data-federation-webapp/src/main/resources

Explanations:
I contribute to the TransitScreen project and we had to add the Metro
North Rail agency to get real time info. The problem is that we need to
set the Accept header to application/x-protobuf to get the proto file
for this agency (see
http://datamine.mta.info/MNR-Train-Time-feed-documentation). Without
setting it, we get an error running OBA: Error updating from
GTFS-realtime data sources. Protocol message end-group tag did not match
expected tag.

Setting the header fixes it.

So for this agency, we can set the headers in the data-sources.xml:
<property name="headers" value="Accept:application/x-protobuf" />

Let me know what you think or let me know if you have any question.