Fix new feature where OBA API predictions are set from TimepointPredictionRecords:
- slack time should be included
- if stop sequence is not included in a timepoint prediction record, and we are interested in a trip that loops, we nee
d to make sure that we have as many TPRs as stops and that we are looking at the right one. Otherwise the TPRs cannot b
e disambiguated. The logic for this is a little complex, and another option would be to fall through in this case to th
e original schedule deviation setting (which presumably handles this case correctly).