Fixes for WMATA-203 - consistency between Nextbus API and OBA
OBA discards TripUpdates in the future (ie all the TripUpdates for a given vehicleId that lose out in the tripMoreAppropriate function). Made a change to keep those TripUpdates but still use the tripMoreAppropriate calculation to set trip effective schedule time.

Fixed a bug in ArrivalAndDepartureServiceImpl where the search for the correct TimepointPredictionRecord fails in some circumstances because it uses OBA stop sequence instead of GTFS stop sequence to find the correct record.