fixed HART issue 89 (https://github.com/camsys/hart/issues/89)
The "Remember Me" checkbox on the admin login didn't seem to connect to anything.  The value of the checkbox is always posted as true no matter if the box is selected or not.  Didn't seem worth it to implement at this time so I commented it out in case someone in the future wants to do something with it.