Deletes seemingly unnecessary Metrics controller and JSP page from tdf-webapp
Related to #390

I had originally added a new controller class to transit-data-federation-webapp for the new Metrics endpoint that is being exposed through the oba-api-webapp project, because this seemed to be required. I no longer think that is the case, and I have removed these two files. I am keeping this as a separate commit to document my evolving understanding of how these projects fit together as there isn't any other documentation to speak of.